fix: move deco application code to plugin
